refactor(types): move ToolResult to oumi.core.types.tool_call
ToolResult is a runtime value type for environment step() returns —
co-locating it with ToolDefinition / ToolCall in tool_call.py groups
all tool-related types in one place and removes a config-vs-types
ownership ambiguity. Drops the unused BaseParams inheritance (it was
an artifact of the file's prior location in core.configs.params).
