+++ src/oumi/environments/synthetic_environment.py
@@ -25,10 +25,10 @@ from oumi.core.configs.params.base_params import BaseParams
 from oumi.core.configs.params.environment_params import EnvironmentParams
 from oumi.core.configs.params.tool_params import (
     ToolParams,
-    ToolResult,
     _validate_json_schema_value,
 )
 from oumi.core.registry import register_environment
+from oumi.core.types.tool_call import ToolResult
 from oumi.environments.base_environment import BaseEnvironment
 from oumi.environments.deterministic_tool import DeterministicTool
