+++ tests/unit/environments/test_synthetic_environment.py
@@ -15,10 +15,8 @@
 import pytest
 
 from oumi.core.configs.params.environment_params import EnvironmentParams
-from oumi.core.configs.params.tool_params import (
-    ToolParams,
-    ToolResult,
-)
+from oumi.core.configs.params.tool_params import ToolParams
+from oumi.core.types.tool_call import ToolResult
 from oumi.environments.deterministic_tool import (
     DeterministicTool,
     DeterministicToolOutput,
